requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-filter==23.5
django-debug-toolbar==4.3.0
drf-spectacular==0.27.1
django-select2==8.1.2
python-magic==0.4.27
Pillow==10.3.0
gunicorn==21.2.0
sentry-sdk==1.40.0
